保存时剔除错误参数，地图删除查询关联

+++ DataAdmin/src/main/java/net/wanji/web/controller/business/AtlasController.java
@@ -1,13 +1,18 @@
 package net.wanji.web.controller.business;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import io.swagger.annotations.*;
 import net.wanji.business.common.Constants;
 import net.wanji.business.domain.dto.TjAtlasTreeDto;
 import net.wanji.business.domain.dto.TjAtlasVenueDto;
+import net.wanji.business.entity.InfinteMileScence;
 import net.wanji.business.entity.TjAtlasVenue;
+import net.wanji.business.entity.TjFragmentedSceneDetail;
 import net.wanji.business.exception.BusinessException;
 import net.wanji.business.service.ITjAtlasTreeService;
 import net.wanji.business.service.ITjAtlasVenueService;
+import net.wanji.business.service.InfinteMileScenceService;
+import net.wanji.business.service.TjFragmentedSceneDetailService;
 import net.wanji.common.core.controller.BaseController;
 import net.wanji.common.core.domain.AjaxResult;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,6 +35,11 @@ public class AtlasController extends BaseController {
     @Autowired
     private ITjAtlasVenueService tjAtlasVenueService;//地图模块场地服务
 
+    @Autowired
+    private InfinteMileScenceService infinteMileScenceService;
+
+    @Autowired
+    private TjFragmentedSceneDetailService tjFragmentedSceneDetailService;
 
 
     @ApiOperationSort(1)
@@ -82,6 +92,16 @@ public class AtlasController extends BaseController {
     @ApiOperation(value = "删除测试场地")
     @GetMapping("deleteVenue")
     public AjaxResult deleteVenue(@RequestParam("id") Integer id) throws BusinessException{
+        QueryWrapper<TjFragmentedSceneDetail> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("map_id", id);
+        if (tjFragmentedSceneDetailService.count(queryWrapper) > 0) {
+            return AjaxResult.error("该场地下有片段式场景，请先删除碎片场景");
+        }
+        QueryWrapper<InfinteMileScence> queryWrapper1 = new QueryWrapper<>();
+        queryWrapper1.eq("map_id", id);
+        if (infinteMileScenceService.count(queryWrapper1) > 0) {
+            return AjaxResult.error("该场地下有无限里程场景，请先删除无限里程场景");
+        }
         return tjAtlasVenueService.deleteVenueById(id)
                 ? AjaxResult.success("成功")
                 : AjaxResult.error("失败");
